[TheNextWeek] Refactor texture API. - It now has all the benefits of previous version but more straightforward and more reliable. - We still have procedure-like API but under the hood we use pointers to functions instead of switches.
